- By default boost will now try to fallback to libraries named in less complete ways then specified by the boost project.
This change was originally added to we could build code on Linux with boost installations that 
had incomplete library names.  This has worked well so I am now going to make it the default.  It 
may still have some unforseen side-effects in the future but overall it seems to make the builds 
behave better.